+++ lib/utils/zkUtils.dart
@@ -11,7 +11,7 @@ const String zkDefaultTokenId =
     "wSHV2S4qX9jFsLjQo8r1BsMLH2ZRKsZx6EJd1sbozGPieEC4Jf";
 
 TransactionDetail getFormatFeePayerV2(
-    Map<dynamic, dynamic> zkappCommand, String currentAddress) {
+    dynamic zkappCommand, String currentAddress) {
   var feePayer = zkappCommand['feePayer'];
   String feePayerKey = feePayer['body']['publicKey'].toString();
   if (feePayerKey.toLowerCase() == zkEmptyPublicKey.toLowerCase()) {
@@ -19,7 +19,6 @@ TransactionDetail getFormatFeePayerV2(
   }
   var fee = feePayer['body']['fee'].toString();
   fee = Fmt.balance(fee.toString(), COIN.decimals, maxLength: COIN.decimals);
-
   return TransactionDetail(
     label: "feePayer",
     children: [
@@ -31,7 +30,8 @@ TransactionDetail getFormatFeePayerV2(
 
 List<dynamic> getZkInfo(String zkappCommand, String currentAddress) {
   try {
-    Map<String, dynamic> nextZkCommand = jsonDecode(zkappCommand);
+    dynamic nextZkCommand = jsonDecode(zkappCommand);
+    nextZkCommand = jsonDecode(nextZkCommand);
     var feePayerBody = getFormatFeePayerV2(nextZkCommand, currentAddress);
     var accountUpdateBody = getUpdateBody(nextZkCommand);
     return [feePayerBody.toJson(), accountUpdateBody.toJson()];
@@ -56,11 +56,12 @@ String getZkFee(String zkappCommand) {
     return "";
   }
 }
+
 String getZkMemo(String zkappCommand) {
   try {
     Map<dynamic, dynamic> nextZkCommand = jsonDecode(zkappCommand);
     var memo = nextZkCommand['memo'];
-    if(memo!= null){
+    if (memo != null) {
       memo = bs58Decode(memo);
     }
     return memo;
@@ -69,7 +70,6 @@ String getZkMemo(String zkappCommand) {
   }
 }
 
-
 AccountUpdateInfo getUpdateBody(Map<String, dynamic> zkappCommand) {
   List<dynamic> accountUpdates = zkappCommand['accountUpdates'];
   List<AccountDetail> children = [];
@@ -77,8 +77,7 @@ AccountUpdateInfo getUpdateBody(Map<String, dynamic> zkappCommand) {
   for (int index = 0; index < accountUpdates.length; index++) {
     var accountItemBody = accountUpdates[index]['body'];
     var publicKey = accountItemBody['publicKey'];
-    var tokenId = accountItemBody['tokenId'] ??
-        zkDefaultTokenId;
+    var tokenId = accountItemBody['tokenId'] ?? zkDefaultTokenId;
     var balanceChangeBody = accountItemBody['balanceChange'];
     var balanceChangeOperator =
         balanceChangeBody['sgn'].toLowerCase() == "negative" ? "-" : "+";
@@ -86,40 +85,56 @@ AccountUpdateInfo getUpdateBody(Map<String, dynamic> zkappCommand) {
     var balanceChange = balanceChangeOperator +
         Fmt.balance(balanceChangeBody['magnitude'].toString(), COIN.decimals,
             maxLength: COIN.decimals);
-    var tokenSymbol = tokenId == zkDefaultTokenId
-        ? COIN.coinSymbol
-        : "UNKNOWN";
+    var tokenSymbol = tokenId == zkDefaultTokenId ? COIN.coinSymbol : "UNKNOWN";
 
     if (tokenId != zkDefaultTokenId &&
         accountItemBody.containsKey('update') &&
         accountItemBody['update']['tokenSymbol'] != null) {
       tokenSymbol = accountItemBody['update']['tokenSymbol'];
     }
-
+    List<Detail> tempDetail = [];
+    tempDetail.add(Detail(label: "publicKey", value: Fmt.address(publicKey, pad: 10)));
+    if(tokenId != zkDefaultTokenId){
+      tempDetail.add(Detail(label: "tokenId", value: Fmt.address(tokenId, pad: 10)));
+    }
+    tempDetail.add(Detail(label: "balanceChange", value: "$balanceChange $tokenSymbol"));
     children.add(AccountDetail(
       label: "Account #${index + 1}",
-      children: [
-        Detail(label: "publicKey", value: Fmt.address(publicKey, pad: 10)),
-        Detail(label: "tokenId", value: Fmt.address(tokenId, pad: 10)),
-        Detail(label: "balanceChange", value: "$balanceChange $tokenSymbol"),
-      ],
+      children: tempDetail
     ));
   }
 
   return AccountUpdateInfo(label: "accountUpdates", children: children);
 }
 
-String getContractAddress(String tx) {
+String getContractAddress(String tx, String currentAccount) {
   try {
     if (tx.isEmpty) return "";
-
-    final Map<String, dynamic> realTx = jsonDecode(tx);
-    final List<dynamic> accountUpdates = realTx['accountUpdates'] ?? [];
-
-    for (final update in accountUpdates) {
-      if (update['authorization'] != null &&
-          update['authorization']['proof'] != null) {
-        return update['body']['publicKey'] ?? "";
+    final dynamic realTx = jsonDecode(tx);
+    final dynamic realTxTemp = jsonDecode(realTx);
+    final List<dynamic> accountUpdates = realTxTemp['accountUpdates'] ?? [];
+    final firstZKapp = accountUpdates.firstWhere(
+      (update) =>
+          update['authorization'] != null &&
+          update['authorization']['proof'] != null,
+      orElse: () => null,
+    );
+    if (firstZKapp != null &&
+        firstZKapp['body'] != null &&
+        firstZKapp['body']['publicKey'] is String) {
+      return firstZKapp['body']['publicKey'];
+    }
+    if (firstZKapp == null) {
+      final firstDifferentPublicKey = accountUpdates.firstWhere(
+        (update) =>
+            update['body'] != null &&
+            update['body']['publicKey'] != currentAccount,
+        orElse: () => null,
+      );
+      if (firstDifferentPublicKey != null &&
+          firstDifferentPublicKey['body'] != null &&
+          firstDifferentPublicKey['body']['publicKey'] is String) {
+        return firstDifferentPublicKey['body']['publicKey'];
       }
     }
   } catch (error) {